Formalize output system; add error message for missing SSL certs
