+++ core/functions.py
@@ -2,6 +2,7 @@ import random
 import string
 import uuid
 
+from core import output
 
 def generate_token(size=32):
     return ''.join(random.SystemRandom().choice(
@@ -37,8 +38,7 @@ def store_creds(
                 zip_code
             ))
     except Exception as ex:
-        print('Error saving creds: {}'.format(ex))
-        pass
+        output.exception(ex)
 
 
 def cache_creds(module, username, password):
@@ -47,8 +47,7 @@ def cache_creds(module, username, password):
             if username and password:
                 fh.write('{},{},{}\n'.format(module, username, password))
     except Exception as ex:
-        print(ex)
-        pass
+        output.exception(ex)
 
 
 def reload_creds(seen):
@@ -68,7 +67,7 @@ def reload_creds(seen):
                     city = cl[7]
                     region = cl[8]
                     zip_code = cl[9]
-                    
+
                     add_cred = {
                         'cred_id': cred_id,
                         'module': module,
@@ -84,38 +83,6 @@ def reload_creds(seen):
                     }
                     creds['creds'].append(add_cred)
     except Exception as ex:
-        print(ex)
-        pass
+        output.exception(ex)
 
     return creds
-
-
-def print_banner():
-    print('''
- ..........................................
- ..........................................
- ..........M...............................
- ........N......................... .Z.....
- .......M.......... M++M..........MMMN.....
- ......MMMM......M?======$M .....MMDM......
- .....,...MMD.MM=====IM=====MM  M..........
- ....D .....M=======M..8=======M ..........
- ...M ....M====+===+...DMM==NM===M.........
- ... .....M==+..MM. ..MMMMMM.MM==M.........
- ..Z......M==M..MMM..NMMMMMMN.MM=M.........
- .........M=M MMMMM..MMMMMMMMMMMMM.........
- .........MMMMMMMMMMMMMMMMMMMMMMMM.........
- ....... .MMMMMMMMMMMMMMMMMMMMMMMM.........
- .........MMMM  BLACK HILLS   MMMM.........
- .........MMMM    INFOSEC     MMMM.........
- ..........MMMMMMMMMMMMMMMMMMMMMM .........
- ...........MMMM,M+IOM=MO=MMNMMMMM.........
- ....... M:. ..MMMMMMMMMMMMMM....MM$.......
- ......MM........+MMMMMMMM.........MM .....
- ...................MMMM...................
- ..........................................
- ..........................................
- ..........................................
-
-[*] CredSniper v1 - Mike Felch (@ustayready)
-''')
