+++ core/output.py
@@ -1,0 +1,43 @@
+
+class Colors(object):
+    N = '\033[m' # native
+    R = '\033[31m' # red
+    G = '\033[32m' # green
+    O = '\033[33m' # orange
+    B = '\033[34m' # blue
+
+def exception(message):
+    print("{r}[!]{n} {message}".format(message=message, r=Colors.R, n=Colors.N))
+
+def information(message):
+    print('{b}[*]{n} {message}'.format(b=Colors.B, n=Colors.N, message=message))
+
+
+def print_banner():
+    print('''
+ ..........................................
+ ..........................................
+ ..........M...............................
+ ........N......................... .Z.....
+ .......M.......... M++M..........MMMN.....
+ ......MMMM......M?======$M .....MMDM......
+ .....,...MMD.MM=====IM=====MM  M..........
+ ....D .....M=======M..8=======M ..........
+ ...M ....M====+===+...DMM==NM===M.........
+ ... .....M==+..MM. ..MMMMMM.MM==M.........
+ ..Z......M==M..MMM..NMMMMMMN.MM=M.........
+ .........M=M MMMMM..MMMMMMMMMMMMM.........
+ .........MMMMMMMMMMMMMMMMMMMMMMMM.........
+ ....... .MMMMMMMMMMMMMMMMMMMMMMMM.........
+ .........MMMM  BLACK HILLS   MMMM.........
+ .........MMMM    INFOSEC     MMMM.........
+ ..........MMMMMMMMMMMMMMMMMMMMMM .........
+ ...........MMMM,M+IOM=MO=MMNMMMMM.........
+ ....... M:. ..MMMMMMMMMMMMMM....MM$.......
+ ......MM........+MMMMMMMM.........MM .....
+ ...................MMMM...................
+ ..........................................
+ ..........................................
+ ..........................................
+ ''')
+    information('CredSniper v1 - Mike Felch (@ustayready)')
